Move codec lookup to rtsp file
